perf: 如果是获取版本信息造成的报错，则不提醒

+++ src/api/index.ts
@@ -31,9 +31,11 @@ class RequestHttp {
         return response.data;
       },
       async (error: AxiosError) => {
+        //如果是获取版本信息造成的报错，则不提醒
+        if (error.config.url?.includes("version")) return;
         if (error.code === "ERR_NETWORK") {
           const reset = confirm(
-            "检测到数据请求失败，请尝试刷新浏览器解决，刷新后如果依旧显示此弹窗，请点击确定清除数据重新下载，如果依旧未解决，请反馈给作者。"
+            "检测到数据请求失败，请尝试点击【取消】后刷新浏览器解决，刷新后如果依旧显示此弹窗，请点击【确定】清除数据重新下载，如果依旧未解决，请反馈给作者。"
           );
           if (reset) {
             localStorage.clear();
